requirements.txt
flask==3.0.0
crewai==0.28.8
crewai[tools]==0.28.8
groq==0.8.0
python-dotenv==1.0.0
pydantic==2.5.3
beautifulsoup4==4.12.2
requests==2.31.0
dnspython==2.4.2
email-validator==2.1.0
PyYAML==6.0.1
markdown==3.5.2
